backend/requirements.txt
fastapi==0.118.0
uvicorn==0.38.0
sqlmodel==0.0.38
openai==2.30.0
python-dotenv==1.0.1
slowapi==0.1.9
pytest==9.0.2
httpx==0.28.1
